order-history

+++ nhom10/src/main/java/com/haui/service/UserService.java
@@ -36,11 +36,11 @@ public class UserService {
         return userRepository.findAll(spec, pageable);
     }
 
-    public User findById(Integer id) {
+    public User findById(long id) {
         return userRepository.findById(id).get();
     }
 
-    public void delete(Integer id) {
+    public void delete(long id) {
         userRepository.deleteById(id);
     }
